+++ src/main/java/com/eureka/mindbloom/member/dto/SignUpResponse.java
@@ -1,4 +1,4 @@
-package com.eureka.mindbloom.auth.dto;
+package com.eureka.mindbloom.member.dto;
 
 import lombok.Builder;
 import lombok.Getter;
