+++ src/main/java/com/eureka/mindbloom/member/service/MemberService.java
@@ -1,7 +1,7 @@
 package com.eureka.mindbloom.member.service;
 
-import com.eureka.mindbloom.auth.dto.SignUpResponse;
-import com.eureka.mindbloom.auth.dto.SignUpRequest;
+import com.eureka.mindbloom.member.dto.SignUpResponse;
+import com.eureka.mindbloom.member.dto.SignUpRequest;
 
 public interface MemberService {
     SignUpResponse signUp(SignUpRequest signUpRequestDto);
